feat: implement Kanban board components with drag-and-drop functionality and enhance task management
